pyproject.toml
[tool.poetry]
name = "FastID"
version = "0.0.4"
description = "Single Sign-On (SSO) Server"
authors = ["Kostya Ten <kostya@yandex.com>"]
readme = "README.md"

include = [
   { path = 'migrations', format = ["sdist", "wheel"] },
   { path = 'alembic.ini', format = ["sdist", "wheel"] }
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.104.1"}
httpx = {extras = ["http2"], version = "^0.25.1"}
uvicorn = "^0.24.0.post1"
asgi-lifespan = "^2.1.0"
python-json-logger = "^2.0.7"
opentelemetry-exporter-otlp-proto-grpc = "^1.21.0"
sqlalchemy = {extras = ["asyncio", "aiosqlite"], version = "^2.0.23"}
asyncpg = "^0.29.0"
prometheus-fastapi-instrumentator = "^6.1.0"
phonenumbers = "^8.13.24"
alembic = "^1.12.0"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"
aiosmtplib = "^3.0.1"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
greenlet = "^3.0.1"
asyncclick = "^8.1.3.4"
async-lru = "^2.0.4"

[tool.poetry.group.dev.dependencies]
mypy = "^1.6.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"
pytest-env = "^1.0.1"
coverage = {extras = ["toml"], version = "^7.3.2"}
pytest-httpx = "^0.26.0"
sqlalchemy = {extras = ["mypy"], version = "^2.0.21"}
setuptools = "^68.2.2"
black = "^23.9.1"
isort = "^5.12.0"
ruff = "^0.1.5"
pytest-mock = "^3.11.1"
pre-commit = "^3.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
user = "fastid.command.user:create"

[tool.mypy]
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true

[tool.pytest.ini_options]
minversion = "6.0"
asyncio_mode = "auto"
testpaths = [
    "tests",
]
env = [
    "PYTEST = True",
    "captcha = recaptcha",
    "captcha_usage = signin",
    "recaptcha_site_key = 6LeIxAcTAAAAAJcZVRqyHh71UMIEGNQ_MXjiZKhI",
    "recaptcha_secret_key = 6LeIxAcTAAAAAGG-vFI1TnRWxMZNFuojJ4WifJWe",
    "cors_enable = True",
]
filterwarnings = [
    "ignore::DeprecationWarning:pkg_resources.*",
    "ignore::DeprecationWarning:google.rpc.*",
    "ignore:.*custom tp_new.*in Python 3.14.*:DeprecationWarning",
    "ignore:datetime.datetime.utcfromtimestamp.*:DeprecationWarning",
    "ignore:datetime.datetime.utcnow.*:DeprecationWarning",
]

[tool.black]
line-length = 120
target-version = ["py312"]
skip-string-normalization = true

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    "venv",
    ".nox",
    ".mypy_cache",
    ".data",
]

#per-file-ignores = {"__init__.py" = ["F401"]}

select = [
    "I", "N", "ASYNC", "B", "A", "COM", "C4", "DTZ", "T10", "EXE", "FA", "ISC",
    "ICN", "PIE", "Q", "RSE", "RET", "SLF", "TD", "FIX", "F",
]

[tool.ruff.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "single"
multiline-quotes = "double"
